Refractored redux folder.
