+++ src/containers/home/index.js
@@ -7,7 +7,7 @@ import {
   incrementAsync,
   decrement,
   decrementAsync
-} from '../../modules/counter'
+} from '../../redux/action-creators'
 
 const Home = props => (
   <div>
