+++ src/redux/action-names/index.js
@@ -1,0 +1,8 @@
+const actions = {
+    INCREMENT_REQUESTED: 'counter/INCREMENT_REQUESTED',
+    INCREMENT: 'counter/INCREMENT',
+    DECREMENT_REQUESTED: 'counter/DECREMENT_REQUESTED',
+    DECREMENT: 'counter/DECREMENT',
+}
+
+export default actions;
