+++ src/redux/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux'
-import counter from './counter'
+import counter from './reducers/counter'
 
 export default combineReducers({
   counter
